@@ -11,14 +11,17 @@ const apiUrl = 'https://api.angularbootcamp.com';
   providedIn: 'root'
 })
 export class EmployeeLoader {
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 
   getList(): Observable<Employee[]> {
-    return this.http.get<Employee[]>(apiUrl + '/employees')
+    return this.http
+      .get<Employee[]>(apiUrl + '/employees')
       .pipe(map(longList => longList.slice(0, 9)));
   }
 
   getDetails(employeeId: number): Observable<Employee> {
-    return this.http.get<Employee>(`${apiUrl}/employees/${employeeId}`);
+    return this.http.get<Employee>(
+      `${apiUrl}/employees/${employeeId}`
+    );
   }
 }
